newApp/django19/requirements.txt
defusedxml==0.4.1
Django==1.9.7
django-tastypie==0.13.3
lxml==3.7.2
pkg-resources==0.0.0
python-dateutil==2.6.0
python-mimeparse==1.6.0
six==1.10.0
